requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.1
langchain>=0.1.14
langchain-core>=0.1.35
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-docx>=1.1.0
chardet>=5.2.0
torch>=2.0.1
langchain-community